refactor(web): 컴포넌트 분리 #261
